@@ -20,41 +20,39 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "~7.2.15",
-    "@angular/core": "~7.2.15",
-    "@angular/fire": "~5.1.3",
-    "@angular/forms": "~7.2.15",
-    "@angular/http": "~7.2.15",
-    "@angular/platform-browser": "~7.2.15",
-    "@angular/platform-browser-dynamic": "~7.2.15",
-    "@angular/pwa": "~0.13.9",
-    "@angular/router": "~7.2.15",
-    "@angular/service-worker": "~7.2.15",
+    "@angular/common": "~8.0.0",
+    "@angular/core": "~8.0.0",
+    "@angular/fire": "~5.2.1",
+    "@angular/forms": "~8.0.0",
+    "@angular/platform-browser": "~8.0.0",
+    "@angular/platform-browser-dynamic": "~8.0.0",
+    "@angular/pwa": "~0.800.1",
+    "@angular/router": "~8.0.0",
+    "@angular/service-worker": "~8.0.0",
     "@ionic/angular": "~4.4.2",
-    "core-js": "~2.6.8",
     "date-fns": "~1.30.1",
-    "firebase": "~6.0.2",
+    "firebase": "~6.1.0",
     "rxjs": "~6.5.2",
-    "zone.js": "~0.8.29"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "~0.13.9",
-    "@angular-devkit/build-angular": "~0.13.9",
-    "@angular-devkit/core": "~7.3.9",
-    "@angular-devkit/schematics": "~7.3.9",
-    "@angular/cli": "~7.3.9",
-    "@angular/compiler": "~7.2.15",
-    "@angular/compiler-cli": "~7.2.15",
-    "@angular/language-service": "~7.2.15",
+    "@angular-devkit/architect": "~0.800.1",
+    "@angular-devkit/build-angular": "~0.800.1",
+    "@angular-devkit/core": "~8.0.1",
+    "@angular-devkit/schematics": "~8.0.1",
+    "@angular/cli": "~8.0.1",
+    "@angular/compiler": "~8.0.0",
+    "@angular/compiler-cli": "~8.0.0",
+    "@angular/language-service": "~8.0.0",
     "@ionic/angular-toolkit": "~1.5.1",
-    "@ionic/lab": "~1.0.24",
-    "@types/jasmine": "~3.3.12",
+    "@ionic/lab": "~2.0.0",
+    "@types/jasmine": "~3.3.13",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~12.0.2",
+    "@types/node": "~12.0.4",
     "adj-noun": "~0.1.1",
-    "axios": "~0.18.0",
+    "axios": "~0.19.0",
     "chalk": "~2.4.2",
-    "codelyzer": "~5.0.1",
+    "codelyzer": "~5.1.0",
     "conventional-changelog": "~3.1.8",
     "execa": "~1.0.0",
     "inquirer": "~6.3.1",
@@ -68,10 +66,10 @@
     "listr": "~0.14.3",
     "protractor": "~5.4.2",
     "rimraf": "~2.6.3",
-    "semver": "~6.1.0",
-    "ts-node": "~8.1.0",
-    "tslint": "~5.16.0",
-    "typescript": "~3.2.4",
+    "semver": "~6.1.1",
+    "ts-node": "~8.2.0",
+    "tslint": "~5.17.0",
+    "typescript": "~3.4.5",
     "webpack-bundle-analyzer": "~3.3.2"
   }
 }
