@@ -3,6 +3,7 @@ import { CanActivate } from '@angular/router';
 import { NavController } from '@ionic/angular';
 
 import { AngularFireAuth } from '@angular/fire/auth';
+import { Subscription } from 'rxjs';
 
 @Injectable({
   providedIn: 'root'
@@ -24,7 +25,8 @@ export class AuthGuardService implements CanActivate {
 
   private getUser(): Promise<any> {
     return new Promise(resolve => {
-      const s = this.afAuth.user.subscribe(u => {
+      let s: Subscription = null;
+      s = this.afAuth.user.subscribe(u => {
         resolve(u);
         if (s) {
           s.unsubscribe();
